Moving methods around from MutableDataset to BaseDataset, some refactors to shoji payloads and class improvements for VariablesMixin
