+++ scrunch/datasets.py
@@ -1,13 +1,14 @@
 import collections
+import datetime
 import json
 import logging
 import re
 import os
 import sys
+import six
 
 import pandas as pd
 import pycrunch
-import six
 from pycrunch.exporting import export_dataset
 
 from scrunch.categories import CategoryList
@@ -15,16 +16,20 @@ from scrunch.exceptions import (InvalidPathError, AuthenticationError,
                                 InvalidReferenceError, OrderUpdateError)
 from scrunch.expressions import parse_expr, prettify, process_expr
 from scrunch.helpers import (ReadOnly, _validate_category_rules,
-                             download_file)
+                             download_file, shoji_wrapper,
+                             abs_url, case_expr, subvar_alias)
 from scrunch.subentity import Deck, Filter, Multitable
+from scrunch.variables import (combinations_from_map, combine_categories_expr,
+                               combine_responses_expr, responses_from_map)
 
+_MR_TYPE = 'multiple_response'
 
 if six.PY2:  # pragma: no cover
-    from urlparse import urlsplit, urljoin
+    from urlparse import urljoin
     import ConfigParser as configparser
 else:
     import configparser
-    from urllib.parse import urlsplit, urljoin
+    from urllib.parse import urljoin
 
 
 LOG = logging.getLogger('scrunch')
@@ -113,13 +118,14 @@ def _get_dataset(dataset, connection=None, editor=False, streaming=False):
                 "Authenticate first with scrunch.connect() or by providing "
                 "config/environment variables")
     root = connection
+    root_datasets = root.datasets
     # search by dataset name
     try:
-        shoji_ds = root.datasets.by('name')[dataset].entity
+        shoji_ds = root_datasets.by('name')[dataset].entity
     except KeyError:
         # search by dataset id
         try:
-            shoji_ds = root.datasets.by('id')[dataset].entity
+            shoji_ds = root_datasets.by('id')[dataset].entity
         except KeyError:
             # search by id on any project
             try:
@@ -163,9 +169,9 @@ def get_user(user, connection=None):
     if connection is None:
         connection = _get_connection()
         if not connection:
-            raise AttributeError(
-                "Authenticate first with scrunch.connect() or by providing "
-                "config/environment variables")
+            raise AuthenticationError(
+                "Unable to find crunch session, crunch.ini file or \
+                environment variables.")
     try:
         ret = connection.users.by('email')[user].entity
     except KeyError:
@@ -946,17 +952,15 @@ class DatasetVariablesMixin(collections.Mapping):
                     self.name, item))
         return Variable(variable, self)
 
+    def _set_catalog(self):
+        self._catalog = self.resource.variables
+
     def _reload_variables(self):
         """
         Helper that takes care of updating self._vars on init and
         whenever the dataset adds a variable
         """
-        # this try except block is needed to update the 
-        # subvar or var catalogs once the Entity is modified
-        try:
-            self._catalog = self.resource.variables
-        except AttributeError:
-            self._catalog = self.resource.subvariables
+        self._set_catalog()
         self._vars = self._catalog.index.items()
 
     def __iter__(self):
@@ -1040,49 +1044,19 @@ class BaseDataset(ReadOnly, DatasetVariablesMixin):
     @editor.setter
     def editor(self, _):
         # Protect the `editor` from external modifications.
-        raise TypeError(
-            'Unsupported operation on the editor property'
-        )
+        raise TypeError('Unsupported operation on the editor property')
 
     def change_editor(self, user):
         """
         Change the current editor of the Crunch dataset.
 
         :param user:
-            The email address, the crunch url or a User instance of the user
-            who should be set as the new current editor of the given dataset.
-
-        :returns: None
+            The email or a User instance of the user who should 
+            be set as the new current editor of the given dataset.
         """
-        def _host_from_url(url):
-            resolved = urlsplit(url)
-            return resolved.hostname
-
-        def _to_url(email):
-            api_users = 'https://{}/api/users/'.format(
-                _host_from_url(self.resource.self)
-            )
-            user_url = None
-
-            users = self.resource.session.get(api_users).payload['index']
-
-            for url, user in six.iteritems(users):
-                if user['email'] == email:
-                    user_url = url
-                    self.resource.patch({'current_editor': url})
-                    break
-            assert user_url is not None, 'Unable to resolve user url'
-
-            return user_url
-
-        def _is_url(u):
-            return u.startswith('https://') or u.startswith('http://')
-
-        if isinstance(user, User):
-            user = user.url
-        user_url = user if _is_url(user) else _to_url(user)
-
-        self.resource.patch({'current_editor': user_url})
+        if not isinstance(user, User):
+            user = get_user(user)
+        self.resource.patch({'current_editor': user.url})
         self.resource.refresh()
 
     @property
@@ -1230,6 +1204,313 @@ class BaseDataset(ReadOnly, DatasetVariablesMixin):
             )
             self._settings = None
 
+    def edit(self, **kwargs):
+        """
+        Edit main Dataset attributes
+        """
+        for key in kwargs:
+            if key not in self._MUTABLE_ATTRIBUTES:
+                raise AttributeError("Can't edit attibute %s of variable %s" % (
+                    key, self.name
+                ))
+            if key in ['start_date', 'end_date'] and \
+                    (isinstance(kwargs[key], datetime.date) or
+                     isinstance(kwargs[key], datetime.datetime)
+                     ):
+                kwargs[key] = kwargs[key].isoformat()
+
+        return self.resource.edit(**kwargs)
+
+    def create_single_response(self, categories, name, alias, description='',
+                               missing=True, notes=''):
+        """
+        Creates a categorical variable deriving from other variables.
+        Uses Crunch's `case` function.
+        """
+        cases = []
+        for cat in categories:
+            cases.append(cat.pop('case'))
+
+        if not hasattr(self.resource, 'variables'):
+            self.resource.refresh()
+
+        args = [{
+            'column': [c['id'] for c in categories],
+            'type': {
+                'value': {
+                    'class': 'categorical',
+                    'categories': categories}}}]
+
+        for cat in args[0]['type']['value']['categories']:
+            cat.setdefault('missing', False)
+
+        if missing:
+            args[0]['column'].append(-1)
+            args[0]['type']['value']['categories'].append(dict(
+                id=-1,
+                name='No Data',
+                numeric_value=None,
+                missing=True))
+
+        more_args = []
+        for case in cases:
+            more_args.append(parse_expr(case))
+
+        more_args = process_expr(more_args, self.resource)
+
+        expr = dict(function='case', args=args + more_args)
+
+        payload = shoji_wrapper(dict(
+            alias=alias,
+            name=name,
+            expr=expr,
+            description=description,
+            notes=notes))
+
+        new_var = self.resource.variables.create(payload)
+        # needed to update the variables collection
+        self._reload_variables()
+        # return the variable instance
+        return self[new_var['body']['alias']]
+
+    def create_multiple_response(self, responses, name, alias, description='', notes=''):
+        """
+        Creates a Multiple response (array) using a set of rules for each
+        of the responses(subvariables).
+        """
+        responses_map = collections.OrderedDict()
+        responses_map_ids = []
+        for resp in responses:
+            case = resp['case']
+            if isinstance(case, six.string_types):
+                case = process_expr(parse_expr(case), self.resource)
+
+            resp_id = '%04d' % resp['id']
+            responses_map_ids.append(resp_id)
+            responses_map[resp_id] = case_expr(
+                case,
+                name=resp['name'],
+                alias='%s_%d' % (alias, resp['id'])
+            )
+
+        payload = shoji_wrapper({
+            'name': name,
+            'alias': alias,
+            'description': description,
+            'notes': notes,
+            'derivation': {
+                'function': 'array',
+                'args': [{
+                    'function': 'select',
+                    'args': [
+                        {'map': responses_map},
+                        {'value': responses_map_ids}
+                    ]
+                }]
+            }
+        })
+
+        new_var = self.resource.variables.create(payload)
+        # needed to update the variables collection
+        self._reload_variables()
+        # return an instance of Variable
+        return self[new_var['body']['alias']]
+
+    def create_numeric(self, alias, name, derivation, description='', notes=''):
+        """
+        Used to create new numeric variables using Crunchs's derived expressions
+        """
+        expr = parse_expr(derivation)
+
+        if not hasattr(self.resource, 'variables'):
+            self.resource.refresh()
+
+        payload = shoji_wrapper(dict(
+            alias=alias,
+            name=name,
+            derivation=expr,
+            description=description,
+            notes=notes
+        ))
+
+        self.resource.variables.create(payload)
+        # needed to update the variables collection
+        self._reload_variables()
+        # return the variable instance
+        return self[alias]
+
+    def create_categorical(self, categories, alias, name, multiple,
+                           description='', notes=''):
+        """
+        Used to create new categorical variables using Crunchs's `case`
+        function
+
+        Will create either categorical variables or multiple response depending
+        on the `multiple` parameter.
+        """
+        if multiple:
+            return self.create_multiple_response(
+                categories, alias=alias, name=name, description=description,
+                notes=notes)
+        else:
+            return self.create_single_response(
+                categories, alias=alias, name=name, description=description,
+                notes=notes)
+
+    def copy_variable(self, variable, name, alias):
+        _subvar_alias = re.compile(r'.+_(\d+)$')
+
+        def subrefs(_variable, _alias):
+            # In the case of MR variables, we want the copies' subvariables
+            # to have their aliases in the same pattern and order that the
+            # parent's are, that is `parent_alias_#`.
+            _subreferences = []
+            for subvar in _variable.resource.subvariables.index.values():
+                sv_alias = subvar['alias']
+                match = _subvar_alias.match(sv_alias)
+                if match:  # Does this var have the subvar pattern?
+                    suffix = int(match.groups()[0], 10)  # Keep the position
+                    sv_alias = subvar_alias(_alias, suffix)
+
+                _subreferences.append({
+                    'name': subvar['name'],
+                    'alias': sv_alias
+                })
+            return _subreferences
+
+        if variable.resource.body.get('derivation'):
+            # We are dealing with a derived variable, we want the derivation
+            # to be executed again instead of doing a `copy_variable`
+            derivation = abs_url(variable.resource.body['derivation'],
+                                 variable.resource.self)
+            derivation.pop('references', None)
+            payload = shoji_wrapper({
+                'name': name,
+                'alias': alias,
+                'derivation': derivation})
+
+            if variable.type == _MR_TYPE:
+                # We are re-executing a multiple_response derivation.
+                # We need to update the complex `array` function expression
+                # to contain the new suffixed aliases. Given that the map is
+                # unordered, we have to iterated and find a name match.
+                subvars = payload['body']['derivation']['args'][0]['args'][0]['map']
+                subreferences = subrefs(variable, alias)
+                for subref in subreferences:
+                    for subvar_pos in subvars:
+                        subvar = subvars[subvar_pos]
+                        if subvar['references']['name'] == subref['name']:
+                            subvar['references']['alias'] = subref['alias']
+                            break
+        else:
+            payload = shoji_wrapper({
+                'name': name,
+                'alias': alias,
+                'derivation': {
+                    'function': 'copy_variable',
+                    'args': [{
+                        'variable': variable.resource.self
+                    }]
+                }
+            })
+            if variable.type == _MR_TYPE:
+                subreferences = subrefs(variable, alias)
+                payload['body']['derivation']['references'] = {
+                    'subreferences': subreferences
+                }
+
+        new_var = self.resource.variables.create(payload)
+        # needed to update the variables collection
+        self._reload_variables()
+        # return an instance of Variable
+        return self[new_var['body']['alias']]
+
+    def combine_categories(self, variable, map, categories, missing=None, default=None,
+                           name='', alias='', description=''):
+        if not alias or not name:
+            raise ValueError("Name and alias are required")
+        if variable.type in _MR_TYPE:
+            return self.combine_multiple_response(variable, map, categories, name=name,
+                                                  alias=alias, description=description)
+        else:
+            return self.combine_categorical(variable, map, categories, missing, default,
+                                            name=name, alias=alias, description=description)
+
+    def combine_categorical(self, variable, map, categories=None, missing=None,
+                            default=None, name='', alias='', description=''):
+        """
+        Create a new variable in the given dataset that is a recode
+        of an existing variable
+            map={
+                1: (1, 2),
+                2: 3,
+                3: (4, 5)
+            },
+            default=9,
+            missing=[-1, 9],
+            categories={
+                1: "low",
+                2: "medium",
+                3: "high",
+                9: "no answer"
+            },
+            missing=9
+        """
+        if isinstance(variable, six.string_types):
+            variable = self[variable]
+
+        # TODO: Implement `default` parameter in Crunch API
+        combinations = combinations_from_map(map, categories or {}, missing or [])
+        payload = shoji_wrapper({
+            'name': name,
+            'alias': alias,
+            'description': description,
+            'derivation': combine_categories_expr(
+                variable.resource.self, combinations)
+        })
+        # this returns an entity
+        new_var = self.resource.variables.create(payload)
+        # needed to update the variables collection
+        self._reload_variables()
+        # at this point we are returning a Variable instance
+        return self[new_var['body']['alias']]
+
+    def combine_multiple_response(self, variable, map, categories=None, default=None,
+                                  name='', alias='', description=''):
+        """
+        Creates a new variable in the given dataset that combines existing
+        responses into new categorized ones
+            map={
+                1: 1,
+                2: [2, 3, 4]
+            },
+            categories={
+                1: "online",
+                2: "notonline"
+            }
+        """
+        if isinstance(variable, six.string_types):
+            parent_alias = variable
+            variable = self[variable]
+        else:
+            parent_alias = variable.alias
+
+        # TODO: Implement `default` parameter in Crunch API
+        responses = responses_from_map(variable, map, categories or {}, alias,
+                                       parent_alias)
+        payload = shoji_wrapper({
+            'name': name,
+            'alias': alias,
+            'description': description,
+            'derivation': combine_responses_expr(
+                variable.resource.self, responses)
+        })
+        new_var = self.resource.variables.create(payload)
+        # needed to update the variables collection
+        self._reload_variables()
+        # return an instance of Variable
+        return self[new_var['body']['alias']]
+
     def create_savepoint(self, description):
         """
         Creates a savepoint on the dataset.
@@ -1246,10 +1527,7 @@ class BaseDataset(ReadOnly, DatasetVariablesMixin):
                     " exists.".format(description)
                 )
 
-        self.resource.savepoints.create({
-            'element': 'shoji:entity',
-            'body': {'description': description}
-        })
+        self.resource.savepoints.create(shoji_wrapper({'description': description}))
 
     def load_savepoint(self, description=None):
         """
@@ -1285,14 +1563,133 @@ class BaseDataset(ReadOnly, DatasetVariablesMixin):
                 'user_name'
                 'version'
         """
-        if len(self.resource.savepoints.index) != 0:
+        svpoints = self.resource.savepoints
+        if len(svpoints.index) != 0:
             attribs = [
                 cp[attrib]
-                for url, cp in six.iteritems(self.resource.savepoints.index)
+                for url, cp in six.iteritems(svpoints.index)
             ]
-
             return attribs
         return []
+
+    def create_crunchbox(
+            self, title='', header='', footer='', notes='',
+            filters=None, variables=None, force=False, min_base_size=None,
+            palette=None):
+        """
+        create a new boxdata entity for a CrunchBox.
+
+        NOTE: new boxdata is only created when there is a new combination of
+        where and filter data.
+
+        Args:
+            title       (str): Human friendly identifier
+            notes       (str): Other information relevent for this CrunchBox
+            header      (str): header information for the CrunchBox
+            footer      (str): footer information for the CrunchBox
+            filters    (list): list of filter names or `Filter` instances
+            where      (list): list of variable aliases or `Variable` instances
+                               If `None` all variables will be included.
+            min_base_size (int): min sample size to display values in graph
+            palette     dict : dict of colors as documented at docs.crunch.io
+                i.e.
+                {
+                    "brand": ["#111111", "#222222", "#333333"],
+                    "static_colors": ["#444444", "#555555", "#666666"],
+                    "base": ["#777777", "#888888", "#999999"],
+                    "category_lookup": {
+                        "category name": "#aaaaaa",
+                        "another category:": "bbbbbb"
+                }
+
+        Returns:
+            CrunchBox (instance)
+        """
+
+        if filters:
+            if not isinstance(filters, list):
+                raise TypeError('`filters` argument must be of type `list`')
+
+            # ensure we only have `Filter` instances
+            filters = [
+                f if isinstance(f, Filter) else self.filters[f]
+                for f in filters
+            ]
+
+            if any(not f.is_public
+                    for f in filters):
+                raise ValueError('filters need to be public')
+
+            filters = [
+                {'filter': f.resource.self}
+                for f in filters
+            ]
+
+        if variables:
+            if not isinstance(variables, list):
+                raise TypeError('`variables` argument must be of type `list`')
+
+            # ensure we only have `Variable` Tuples
+            # NOTE: if we want to check if variables are public we would have
+            # to use Variable instances instead of their Tuple representation.
+            # This would cause additional GET's
+            variables = [
+                var.shoji_tuple if isinstance(var, Variable)
+                else self.resource.variables.by('alias')[var]
+                for var in variables
+            ]
+
+            variables = dict(
+                function='select',
+                args=[
+                    {'map': {
+                        v.id: {'variable': v.entity_url}
+                        for v in variables
+                    }}
+                ])
+
+        if not title:
+            title = 'CrunchBox for {}'.format(str(self))
+
+        payload = shoji_wrapper(dict(
+            where=variables,
+            filters=filters,
+            force=force,
+            title=title,
+            notes=notes,
+            header=header,
+            footer=footer)
+        )
+
+        if min_base_size:
+            payload['body'].setdefault('display_settings', {}).update(
+                dict(minBaseSize=dict(value=min_base_size)))
+        if palette:
+            payload['body'].setdefault('display_settings', {}).update(
+                dict(palette=palette))
+
+        # create the boxdata
+        self.resource.boxdata.create(payload)
+
+        # NOTE: the entity from the response is a bit different compared to
+        # others, i.e. no id, no delete method, different entity_url...
+        # For now, return the shoji_tuple from the index
+        for shoji_tuple in self.resource.boxdata.index.values():
+            if shoji_tuple.metadata.title == title:
+                return CrunchBox(shoji_tuple, self)
+
+    def delete_crunchbox(self, **kwargs):
+        """ deletes crunchboxes on matching kwargs """
+        match = False
+        for key in kwargs:
+            if match:
+                break
+            for crunchbox in self.crunchboxes:
+                attr = getattr(crunchbox, key, None)
+                if attr and attr == kwargs[key]:
+                    crunchbox.remove()
+                    match = True
+                    break
 
     def forks_dataframe(self):
         """
@@ -1488,33 +1885,190 @@ class BaseDataset(ReadOnly, DatasetVariablesMixin):
         )
 
     def get_exclusion(self):
-        # make sure there is an expression
-        expr = self.resource.exclusion.get('body').get('expression')
-        if not expr:
+        exclusion = self.resource.exclusion
+        if 'body' not in exclusion:
             return None
-        return prettify(self.resource.exclusion.body.expression, self)
+        expr = exclusion['body'].get('expression')
+        return prettify(expr, self) if expr else None
 
     def add_filter(self, name, expr, public=False):
-        payload = dict(element='shoji:entity',
-                       body=dict(name=name,
-                                 expression=process_expr(parse_expr(expr), self.resource),
-                                 is_public=public))
+        payload = shoji_wrapper(dict(
+            name=name,
+            expression=process_expr(parse_expr(expr), self.resource),
+            is_public=public))
         new_filter = self.resource.filters.create(payload)
         return self.filters[new_filter.body['name']]
 
     def add_deck(self, name, description="", public=False):
-        payload = dict(element='shoji:entity',
-                       body=dict(name=name,
-                                 description=description,
-                                 is_public=public))
+        payload = shoji_wrapper(dict(
+            name=name,
+            description=description,
+            is_public=public))
         new_deck = self.resource.decks.create(payload)
         return self.decks[new_deck.self.split('/')[-2]]
 
+    def fork(self, description=None, name=None, is_published=False,
+             preserve_owner=False, **kwargs):
+        """
+        Create a fork of ds and add virgin savepoint.
 
-class Variable(ReadOnly, DatasetVariablesMixin):
+        :param description: str, default=None
+            If given, the description to be applied to the fork. If not
+            given the description will be copied from ds.
+        :param name: str, default=None
+            If given, the name to be applied to the fork. If not given a
+            default name will be created which numbers the fork based on
+            how many other forks there are on ds.
+        :param is_published: bool, default=False
+            If True, the fork will be visible to viewers of ds. If False it
+            will only be viewable to editors of ds.
+        :param preserve_owner: bool, default=False
+            If True, the owner of the fork will be the same as the parent
+            dataset. If the owner of the parent dataset is a Crunch project,
+            then it will be preserved regardless of this parameter.
+
+        :returns _fork: scrunch.datasets.Dataset
+            The forked dataset.
+        """
+        nforks = len(self.resource.forks.index)
+        if name is None:
+            name = "FORK #{} of {}".format(nforks + 1, self.resource.body.name)
+        if description is None:
+            description = self.resource.body.description
+
+        body = dict(
+            name=name,
+            description=description,
+            is_published=is_published,
+            **kwargs)
+
+        if preserve_owner or '/api/projects/' in self.resource.body.owner:
+            body['owner'] = self.resource.body.owner
+        # not returning a dataset
+        payload = shoji_wrapper(body)
+        _fork = self.resource.forks.create(payload).refresh()
+        return BaseDataset(_fork)
+
+    def merge(self, fork_id=None, autorollback=True):
+        """
+        :param fork_id: str or int
+            can either be the fork id, name or its number as string or int
+
+        :param autorollback: bool, default=True
+            if True the original dataset is rolled back to the previous state
+            in case of a merge conflict.
+            if False the dataset and fork are beeing left 'dirty'
+        """
+        if isinstance(fork_id, int) or (
+                isinstance(fork_id, six.string_types) and
+                fork_id.isdigit()):
+            fork_id = "FORK #{} of {}".format(fork_id, self.resource.body.name)
+
+        elif fork_id is None:
+            raise ValueError('fork id, name or number missing')
+
+        fork_index = self.resource.forks.index
+
+        forks = [f for f in fork_index
+                 if fork_index[f].get('name') == fork_id or
+                 fork_index[f].get('id') == fork_id]
+        if len(forks) == 1:
+            fork_url = forks[0]
+        else:
+            raise ValueError(
+                "Couldn't find a (unique) fork. "
+                "Please try again using its id")
+
+        body = dict(
+            dataset=fork_url,
+            autorollback=autorollback)
+        self.resource.actions.create(body)
+
+    def delete_forks(self):
+        """
+        Deletes all the forks on the dataset. CANNOT BE UNDONE!
+        """
+        for fork in six.itervalues(self.resource.forks.index):
+            fork.entity.delete()
+
+    def create_multitable(self, name, template, is_public=False):
+        """
+        template: List of dictionaries with the following keys
+        {"query": <query>, "transform": <transform>|optional}.
+        A query is a variable or a function on a variable:
+        {"query": bin(birthyr)}
+
+        If transform is specified it must have the form
+        {
+            "query": var_x,
+            "transform": {"categories": [
+                {
+                    "missing": false,  --> default: False
+                    "hide": true,      --> default: True
+                    "id": 1,
+                    "name": "not asked"
+                },
+                {
+                    "missing": false,
+                    "hide": true,
+                    "id": 4,
+                    "name": "skipped"
+                }
+            ]}
+        }
+        """
+        # build template payload
+        parsed_template = []
+        for q in template:
+            # sometimes q is not a dict but simply a string, convert it to a dict
+            if isinstance(q, str):
+                q = {"query": q}
+            as_json = {}
+            parsed_q = process_expr(parse_expr(q['query']), self.resource)
+            # wrap the query in a list of one dict element
+            as_json['query'] = [parsed_q]
+            if 'transform' in q.keys():
+                as_json['transform'] = q['transform']
+            parsed_template.append(as_json)
+        payload = shoji_wrapper(dict(
+            name=name,
+            is_public=is_public,
+            template=parsed_template))
+        new_multi = self.resource.multitables.create(payload)
+        return self.multitables[new_multi.body['name']]
+
+    def import_multitable(self, name, multi):
+        """
+        Copies a multitable from another Dataset into this one:
+        As described at http://docs.crunch.io/#post176
+        :name: Name of the new multitable
+        :multi: Multitable instance to clone into this Dataset
+        """
+        payload = shoji_wrapper(dict(
+            name=name,
+            multitable=multi.resource.self))
+        self.resource.multitables.create(payload)
+        return self.multitables[name]
+
+
+class DatasetSubvariablesMixin(DatasetVariablesMixin):
+
+    def _reload_variables(self):
+        """
+        Helper that takes care of updating self._vars on init and
+        whenever the dataset adds a variable
+        """
+        self._vars = []
+        self._catalog = {}
+        if getattr(self.resource, 'subvariables', None):
+            self._catalog = self.resource.subvariables
+            self._vars = self._catalog.index.items()
+
+
+class Variable(ReadOnly, DatasetSubvariablesMixin):
     """
     A pycrunch.shoji.Entity wrapper that provides variable-specific methods.
-    DatasetVariablesMixin provides for subvariable interactions.
+    DatasetSubvariablesMixin provides for subvariable interactions.
     """
     _MUTABLE_ATTRIBUTES = {'name', 'description',
                            'view', 'notes', 'format'}
@@ -1536,11 +2090,7 @@ class Variable(ReadOnly, DatasetVariablesMixin):
         self._resource = None
         self.url = var_tuple.entity_url
         self.dataset = dataset
-        # for Variables with subvariables
-        try:
-            self._reload_variables()
-        except AttributeError:
-            pass
+        self._reload_variables()
 
     @property
     def resource(self):
@@ -1603,10 +2153,7 @@ class Variable(ReadOnly, DatasetVariablesMixin):
         more_args = process_expr(more_args, self.dataset)
         # epression value building
         expr = dict(function='case', args=args + more_args)
-        payload = dict(
-            element='shoji:entity',
-            body=dict(expr=expr)
-        )
+        payload = shoji_wrapper(dict(expr=expr))
         # patch the variable with the new payload
         resp = self.resource.patch(payload)
         self._reload_variables()
