+++ scrunch/streaming_dataset.py
@@ -1,4 +1,6 @@
+from pycrunch.importing import Importer
 from scrunch.datasets import BaseDataset, _get_dataset
+from scrunch.helpers import shoji_wrapper
 from scrunch.exceptions import InvalidDatasetTypeError
 
 
@@ -22,4 +24,31 @@ class StreamingDataset(BaseDataset):
     A Crunch entity that represents Datasets that are currently
     of the "streaming" class
     """
-    pass
+
+    def stream_rows(self, columns):
+        """
+        Receives a dict with columns of values to add and streams them
+        into the dataset. Client must call .push_rows(n) later or wait until
+        Crunch automatically processes the batch.
+
+        Returns the total of rows streamed
+        """
+        importer = Importer()
+        count = len(list(columns.values())[0])
+        for x in range(count):
+            importer.stream_rows(self.resource,
+                                 {a: columns[a][x] for a in columns})
+        return count
+
+    def push_rows(self, count=None):
+        """
+        Batches in the rows that have been recently streamed. This forces
+        the rows to appear in the dataset instead of waiting for crunch
+        automatic batcher process.
+        """
+        if bool(self.resource.stream.body.pending_messages):
+            self.resource.batches.create(
+                shoji_wrapper({
+                    'stream': count,
+                    'type': 'ldjson'}
+                ))
