+++ scrunch/tests/test_datasets.py
@@ -14,9 +14,10 @@ from pycrunch.elements import JSONObject, ElementSession
 from pycrunch.variables import cast
 
 import scrunch
-from scrunch.datasets import Variable
+from scrunch.datasets import Variable, BaseDataset
 from scrunch.subentity import Filter, Multitable, Deck
 from scrunch.mutable_dataset import MutableDataset
+from scrunch.streaming_dataset import StreamingDataset
 from scrunch.tests.test_categories import EditableMock, TEST_CATEGORIES
 
 
@@ -165,7 +166,7 @@ class TestDatasets(TestDatasetBase, TestCase):
 
     def test_edit_dataset(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
 
         assert ds.name == 'test_dataset_name'
         changes = dict(name='changed')
@@ -220,7 +221,7 @@ class TestDatasets(TestDatasetBase, TestCase):
             },
         }
         ds_mock = self._dataset_mock(variables=variables)
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
 
         ds.resource = mock.MagicMock()
 
@@ -259,7 +260,7 @@ class TestDatasets(TestDatasetBase, TestCase):
 
     def test_create_crunchbox_full(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
 
         call_params = dict(
             title='my title',
@@ -297,7 +298,7 @@ class TestDatasets(TestDatasetBase, TestCase):
 
     def test_create_crunchbox_defaults(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
 
         expected_payload = {
             'element': 'shoji:entity',
@@ -323,7 +324,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
         apply an exclusion filter to a dataset.
         """
         ds_res = self._dataset_mock()
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         var = ds['var1_alias']
 
         # Action!
@@ -353,7 +354,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
         clear (i.e. remove) the exclusion filter from a dataset.
         """
         ds_res = MagicMock()
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         ds.exclude()
 
         ds.resource.session.patch.assert_called_once_with(
@@ -368,7 +369,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_gt(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'var1_alias > 5')
@@ -385,7 +386,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_in(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'var1_alias in [32766]')
@@ -403,7 +404,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_in_multiple(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'var1_alias in (32766, 32767)')
@@ -436,7 +437,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
         }
 
         ds_mock = self._dataset_mock(variables=variables)
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var1 = ds['disposition']
         var2 = ds['exit_status']
 
@@ -483,7 +484,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_any(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'var1_alias.any([32766])')
@@ -507,7 +508,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_not_any(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'not var1_alias.any([32766])')
@@ -536,7 +537,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_any_multiple(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'var1_alias.any([32766, 32767])')
@@ -561,7 +562,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_all(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'var1_alias.all([32767])')
@@ -583,7 +584,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_not_all(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'not var1_alias.all([32767])')
@@ -612,7 +613,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_all_or_all(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'var1_alias.all([1]) or var1_alias.all([2])')
@@ -654,7 +655,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_not_all_or_all(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'not(var1_alias.all([1]) or var1_alias.all([2]))')
@@ -701,7 +702,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_duplicates(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'var1_alias.duplicates()')
@@ -720,7 +721,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_valid(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'valid(var1_alias)')
@@ -739,7 +740,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_not_valid(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'not valid(var1_alias)')
@@ -763,7 +764,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_missing(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'missing(var1_alias)')
@@ -782,7 +783,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_not_missing(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'not missing(var1_alias)')
@@ -806,7 +807,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_equal(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         data = self._exclude_payload(ds, 'var1_alias == 1')
@@ -842,7 +843,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
             )
         }
         ds_mock = self._dataset_mock(variables=variables)
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var1 = ds['disposition']
         var2 = ds['exit_status']
 
@@ -963,7 +964,7 @@ class TestExclusionFilters(TestDatasetBase, TestCase):
 
     def test_dict_expr(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
 
         expr = {
             "args": [
@@ -986,7 +987,7 @@ class TestProtectAttributes(TestDatasetBase, TestCase):
 
     def test_Dataset_attribute_writes(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         assert ds.name == 'test_dataset_name'
 
         with pytest.raises(AttributeError, message=self.error_msg):
@@ -1019,7 +1020,7 @@ class TestProtectAttributes(TestDatasetBase, TestCase):
 
     def test_Variable_attribute_writes(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         with pytest.raises(AttributeError, message=self.error_msg):
@@ -1046,7 +1047,7 @@ class TestProtectAttributes(TestDatasetBase, TestCase):
 class TestVariables(TestDatasetBase, TestCase):
     def test_variable_as_member(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         assert ds.name == self.ds_shoji['body']['name']
         assert ds.id == self.ds_shoji['body']['id']
 
@@ -1060,7 +1061,7 @@ class TestVariables(TestDatasetBase, TestCase):
         with pytest.raises(AttributeError) as err:
             ds.some_variable
         assert str(err.value) == \
-            "'MutableDataset' object has no attribute 'some_variable'"
+            "'StreamingDataset' object has no attribute 'some_variable'"
 
     def test_variable_cast(self):
         variable = MagicMock()
@@ -1082,7 +1083,7 @@ class TestVariables(TestDatasetBase, TestCase):
 
     def test_edit_Variables(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         var = ds['var1_alias']
 
         assert var.name == 'var1_name'
@@ -1120,18 +1121,14 @@ class TestCurrentEditor(TestDatasetBase, TestCase):
     ds_url = 'https://test.crunch.io/api/datasets/123456/'
     user_url = 'https://test.crunch.io/api/users/12345/'
 
-    def test_change_editor(self):
-        ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
-        ds.change_editor(self.user_url)
-
-        ds_mock.patch.assert_called_with({
-            'current_editor': self.user_url
-        })
-
-    def test_change_editor_email(self):
+    @mock.patch('scrunch.datasets.get_user')
+    def test_change_editor_email(self, mocked_get_user):
         sess = MagicMock()
         response = MagicMock()
+        user = MagicMock()
+        user.resource.self = self.user_url
+        user.url = self.user_url
+        mocked_get_user.return_value = user
         response.payload = {
             'index': {
                 self.user_url: {
@@ -1147,7 +1144,7 @@ class TestCurrentEditor(TestDatasetBase, TestCase):
         ds_res = MagicMock(session=sess)
         ds_res.self = self.ds_url
         ds_res.patch = MagicMock()
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         ds.change_editor('jane.doe@crunch.io')
 
         ds_res.patch.assert_called_with({
@@ -1162,7 +1159,7 @@ class TestCurrentOwner(TestDatasetBase, TestCase):
 
     def test_change_owner_exception(self):
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         with pytest.raises(AttributeError) as e:
             ds.change_owner(user=self.user_url, project=self.project_url)
             assert e.message == "Must provide user or project. Not both"
@@ -1174,7 +1171,7 @@ class TestCurrentOwner(TestDatasetBase, TestCase):
         user.url = self.user_url
         mocked_get_user.return_value = user
         ds_mock = self._dataset_mock()
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         ds.change_owner(user=user)
         ds_mock.patch.assert_called_with({
             'owner': self.user_url
@@ -1189,7 +1186,7 @@ class TestSavepoints(TestCase):
         sess = MagicMock()
         ds_res = MagicMock(session=sess)
         ds_res.savepoints = MagicMock()
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         ds.create_savepoint('savepoint description')
         ds_res.savepoints.create.assert_called_with({
             'element': 'shoji:entity',
@@ -1207,7 +1204,7 @@ class TestSavepoints(TestCase):
                 'description': 'savepoint description'
             }
         }
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         with pytest.raises(KeyError):
             ds.create_savepoint('savepoint description')
 
@@ -1220,7 +1217,7 @@ class TestSavepoints(TestCase):
                 'description': 'savepoint description'
             }
         }
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         with pytest.raises(KeyError):
             ds.create_savepoint('savepoint description')
 
@@ -1229,7 +1226,7 @@ class TestSavepoints(TestCase):
         ds_res = MagicMock(session=sess)
         ds_res.savepoints = MagicMock()
         ds_res.savepoints.index = {}
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         with pytest.raises(KeyError):
             ds.load_savepoint('savepoint')
 
@@ -1248,9 +1245,10 @@ class TestForks(TestCase):
         ds_res = MagicMock(session=sess, body=body)
         ds_res.forks = MagicMock()
         ds_res.forks.index = {}
-        ds = MutableDataset(ds_res)
+        ds = BaseDataset(ds_res)
         f = ds.fork()
         ds_res.forks.create.assert_called_with({
+            'element': 'shoji:entity',
             'body': {
                 'name': 'FORK #1 of ds name',
                 'description': 'ds description',
@@ -1269,9 +1267,17 @@ class TestForks(TestCase):
         ds_res = MagicMock(session=sess, body=body)
         ds_res.forks = MagicMock()
         ds_res.forks.index = {}
-        ds = MutableDataset(ds_res)
+        ds = BaseDataset(ds_res)
         f = ds.fork(preserve_owner=True)
-        f.resource.patch.assert_called_with({'owner': user_id})
+        ds_res.forks.create.assert_called_with({
+            'element': 'shoji:entity',
+            'body': {
+                'name': 'FORK #1 of ds name',
+                'description': 'ds description',
+                'owner': user_id,
+                'is_published': False,
+            }
+        })
 
     def test_fork_preserve_owner_project(self):
         project_id = 'http://test.crunch.io/api/projects/456/'
@@ -1284,9 +1290,17 @@ class TestForks(TestCase):
         ds_res = MagicMock(session=sess, body=body)
         ds_res.forks = MagicMock()
         ds_res.forks.index = {}
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         f = ds.fork()
-        f.resource.patch.assert_called_with({'owner': project_id})
+        ds_res.forks.create.assert_called_with({
+            'element': 'shoji:entity',
+            'body': {
+                'name': 'FORK #1 of ds name',
+                'description': 'ds description',
+                'owner': project_id,
+                'is_published': False,
+            }
+        })
 
     def test_delete_forks(self):
         f1 = MagicMock()
@@ -1301,7 +1315,7 @@ class TestForks(TestCase):
             'abc3': f3
         }
 
-        ds = MutableDataset(ds_res)
+        ds = BaseDataset(ds_res)
         ds.delete_forks()
 
         assert f1.entity.delete.call_count == 1
@@ -1326,7 +1340,7 @@ class TestForks(TestCase):
             'abc1': f1
         }
 
-        ds = MutableDataset(ds_res)
+        ds = BaseDataset(ds_res)
         df = ds.forks_dataframe()
         assert isinstance(df, DataFrame)
         keys = [k for k in df.keys()]
@@ -1341,7 +1355,7 @@ class TestForks(TestCase):
         ds_res.forks = MagicMock()
         ds_res.forks.index = {}
 
-        ds = MutableDataset(ds_res)
+        ds = BaseDataset(ds_res)
         df = ds.forks_dataframe()
 
         assert df is None
@@ -1372,7 +1386,7 @@ class TestForks(TestCase):
                 'id': 'ghi',
             }
         }
-        ds = MutableDataset(ds_res)
+        ds = BaseDataset(ds_res)
 
         expected_call = {
             'dataset': fork1_url,
@@ -1380,22 +1394,22 @@ class TestForks(TestCase):
         }
 
         ds.merge(1)  # number as int
-        ds_res.actions.post.assert_called_once_with(expected_call)
+        ds_res.actions.create.assert_called_once_with(expected_call)
         ds_res.reset_mock()
         ds.merge('1')  # number as str
-        ds_res.actions.post.assert_called_once_with(expected_call)
+        ds_res.actions.create.assert_called_once_with(expected_call)
         ds_res.reset_mock()
         ds.merge('FORK #1 of ds name')  # name
-        ds_res.actions.post.assert_called_once_with(expected_call)
+        ds_res.actions.create.assert_called_once_with(expected_call)
         ds_res.reset_mock()
         ds.merge('abc')  # id
-        ds_res.actions.post.assert_called_once_with(expected_call)
+        ds_res.actions.create.assert_called_once_with(expected_call)
         ds_res.reset_mock()
 
         # test autorollback=False
         expected_call['autorollback'] = False
         ds.merge(1, autorollback=False)  # number as int
-        ds_res.actions.post.assert_called_once_with(expected_call)
+        ds_res.actions.create.assert_called_once_with(expected_call)
         ds_res.reset_mock()
 
         # ValueError if no unique fork could be found
@@ -1408,12 +1422,12 @@ class TestForks(TestCase):
         expected_call['autorollback'] = True
 
         ds.merge('def')
-        ds_res.actions.post.assert_called_once_with(expected_call)
+        ds_res.actions.create.assert_called_once_with(expected_call)
         ds_res.reset_mock()
 
         expected_call['dataset'] = fork3_url
         ds.merge('ghi')
-        ds_res.actions.post.assert_called_once_with(expected_call)
+        ds_res.actions.create.assert_called_once_with(expected_call)
 
 
 class TestRecode(TestDatasetBase):
@@ -1456,7 +1470,7 @@ class TestRecode(TestDatasetBase):
             },
         }
         ds_mock = self._dataset_mock(variables=variables)
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
 
         responses = [
             {'id': 1, 'name': 'Facebook', 'case': 'var_a > 5'},
@@ -1578,7 +1592,7 @@ class TestRecode(TestDatasetBase):
             },
         }
         ds_mock = self._dataset_mock(variables=variables)
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
 
         responses = [
             {'id': 1, 'name': 'Facebook', 'case': 'var_a > 5'},
@@ -1717,7 +1731,7 @@ class TestCopyVariable(TestCase):
         var_res = mock.MagicMock()
         var_res.entity.body = {'type': 'numeric'}
         var_res.entity.self = '/variable/url/'
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         var = Variable(var_res, ds_res)
         ds.copy_variable(var, name='copy', alias='copy')
         ds_res.variables.create.assert_called_with({
@@ -1753,7 +1767,7 @@ class TestCopyVariable(TestCase):
                 }]
             }}
         var_res.entity.self = '/variable/url/'
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         var = Variable(var_res, ds_res)
         ds.copy_variable(var, name='copy', alias='copy')
         ds_res.variables.create.assert_called_with({
@@ -1969,7 +1983,7 @@ class TestHierarchicalOrder(TestCase):
         ds_resource.variables.orders.hier = '%svariables/hier/' % self.ds_url
         ds_resource.variables.by.return_value = variables
         ds_resource.session.get.side_effect = _session_get
-        self.ds = MutableDataset(ds_resource)
+        self.ds = StreamingDataset(ds_resource)
         self.ds._revision = 'one'
         self.ds._hier_calls = 0
 
@@ -3551,7 +3565,7 @@ class TestDatasetSettings(TestCase):
         ds_resource.self = self.ds_url
         ds_resource.fragments.settings = '%ssettings/' % self.ds_url
         ds_resource.session.get.side_effect = _session_get
-        self.ds = MutableDataset(ds_resource)
+        self.ds = StreamingDataset(ds_resource)
 
     def test_settings_are_displayed_as_dict_obj(self):
         ds = self.ds
@@ -3702,7 +3716,7 @@ class TestDatasetExport(TestCase):
     def setUp(self):
         ds_resource = mock.MagicMock()
         ds_resource.self = self.ds_url
-        self.ds = MutableDataset(ds_resource)
+        self.ds = StreamingDataset(ds_resource)
 
     def test_basic_csv_export(self, export_ds_mock, dl_file_mock):
         ds = self.ds
@@ -3839,12 +3853,12 @@ class TestVariableIterator(TestDatasetBase):
 
     def test_ds_keys(self):
         ds_mock = self._dataset_mock(variables=self.variables)
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         assert isinstance(ds.keys(), list)
 
     def test_ds_values(self):
         ds_mock = self._dataset_mock(variables=self.variables)
-        ds = MutableDataset(ds_mock)
+        ds = StreamingDataset(ds_mock)
         assert isinstance(ds.values(), list)
 
     def test_subvar_order(self):
@@ -3883,11 +3897,11 @@ class TestVariableIterator(TestDatasetBase):
         var_tuple = mock.MagicMock()
         var_tuple.entity.__getitem__.side_effect = getitem
         # mock it as variables to mimick a subvariables catalog
-        var_tuple.entity.variables.index = subvars
+        var_tuple.entity.subvariables.index = subvars
 
         v = Variable(var_tuple=var_tuple, dataset=ds)
 
-        all_ids = [sv[0] for sv in v]
+        all_ids = [sv[1]['id'] for sv in v]
         assert sorted(all_ids) == ['0001', '0002', '0003', '0004']
 
 
@@ -3915,10 +3929,10 @@ class TestFilter(TestDatasetBase, TestCase):
         }
     }
 
-    @mock.patch('scrunch.mutable_dataset.MutableDataset.filters')
+    @mock.patch('scrunch.streaming_dataset.StreamingDataset.filters')
     def test_add_filter(self, filters):
         ds_res = self._dataset_mock()
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         var = ds['var1_alias']
 
         ds.add_filter(name='filter', expr='var1_alias != 0')
@@ -3966,10 +3980,10 @@ class TestDeck(TestDatasetBase, TestCase):
         }
     }
 
-    @mock.patch('scrunch.mutable_dataset.MutableDataset.decks')
+    @mock.patch('scrunch.streaming_dataset.StreamingDataset.decks')
     def test_add_deck(self, decks):
         ds_res = self._dataset_mock()
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
 
         ds.add_deck(name='mydeck', description='description')
 
@@ -3985,7 +3999,7 @@ class TestDeck(TestDatasetBase, TestCase):
 
     def test_deck_accessor(self):
         ds_res = self._dataset_mock()
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
 
         deck = EditableMock(entity=self._deck)
         Deck(deck)
@@ -4026,10 +4040,10 @@ class TestMultitable(TestDatasetBase, TestCase):
         }
     }
 
-    @mock.patch('scrunch.mutable_dataset.MutableDataset.multitables')
+    @mock.patch('scrunch.streaming_dataset.StreamingDataset.multitables')
     def test_add_multitable(self, multitables):
         ds_res = self._dataset_mock()
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         ds.create_multitable(name='mymulti', template=['var1_alias'])
         expected_payload = {
             'element': 'shoji:entity',
@@ -4049,14 +4063,14 @@ class TestMultitable(TestDatasetBase, TestCase):
 
     def test_multitable_accessor(self):
         ds_res = self._dataset_mock()
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         mt = EditableMock(entity=self._multitable)
         Multitable(mt, ds)
         assert type(ds.multitables) == dict
 
     def test_edit_multitable(self):
         ds_res = self._dataset_mock()
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         mt = EditableMock(entity=self._multitable)
         mockmulti = Multitable(mt, ds)
         with pytest.raises(AttributeError):
@@ -4065,17 +4079,17 @@ class TestMultitable(TestDatasetBase, TestCase):
 
     def test_multitable_class(self):
         ds_res = self._dataset_mock()
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         mt = MagicMock(entity=self._multitable)
         mockmulti = Multitable(mt, ds)
         assert mockmulti
 
     def test_multitable_import(self):
         ds_res = self._dataset_mock()
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         mt = EditableMock(entity=self._multitable)
         mockmulti = Multitable(mt, ds)
-        ds_2 = MutableDataset(ds_res)
+        ds_2 = StreamingDataset(ds_res)
         expected_payload = {
             'element': 'shoji:entity',
             'body': {
@@ -4089,7 +4103,7 @@ class TestMultitable(TestDatasetBase, TestCase):
 
     def test_export_tabbook(self):
         ds_res = self._dataset_mock()
-        ds = MutableDataset(ds_res)
+        ds = StreamingDataset(ds_res)
         mt = EditableMock(entity=self._multitable)
         mockmulti = Multitable(mt, ds)
         expected_payload = {
